inference: Don't try to infer optimized opaque_closure (#56557)
We don't have frontend syntax for it, but there is a use case for having
`:new_opaque_closure` take an OC constructed from an optimized
OpaqueClosure (and just replacing the capture environment). In this
case, there is nothing inference can do to introspect into the opaque
closure, so it just needs to bail out early.