Don't throw an error in raw! if the stream is closed
This was added in #12568 to protect against a segfault after `close(stdin)`.
However, the API is not great, because the stdin closing is an asynchronous
event, so there isn't really any way to use this API without inccurring
an error. Further, it already returns an error code of whether or not
the action suceeded, and it's bad practice to have two ways for an operation
to fail. Remove the error check and handle a closed stream gracefully returning
an EOF error. In all users in Base, this EOF error is ignored, but we will
gracefully check for EOF later and shut down the REPL, which is the desired
behavior.

Fixes https://github.com/timholy/Revise.jl/issues/859