make `choosetests` find the test files for devved stdlibs (#56558)
I want to move out LinearAlgebra into its own repository but I still
want to be able to run its tests in parallel. The easiest would be to be
able to use `Base.runtests` but right now it hard codes the path to the
stdlib folder. Instead, use the loading mechanism to look up where the
stdlib of the active project actively resides.