precompile: improve precision of extension dependencies
This change is a slimmed-down version of #55910, without any of the
re-factoring.

This improves the parallelism of the pre-compile job and fixes a
correctness bug, where unconditionally injecting a dependency after its
parent could be in conflict with dependencies in between triggers,
causing false extension cycles.

For example adding `AMDGPU.jl` and `Tracker.jl` in your project reports
a false cycle (without this change).