support passing a specific Method to invoke
The main purpose of this is integrating with external compilers using
overlay method tables, where the Method might need to come from a source
other than regular dispatch. Note that this is generally much less well
optimized at runtime than generic dispatch, so this shouldn't be treated
as expecting to give a performance boost in any dynamic cases.

Trivial examples:

    julia> let m = which(+, (Int, Int))
           @eval f(i, j) = invoke(+, $m, i, j)
           end
    julia> f(2,2)
    4

    julia> let m = which(Core.kwcall, (@NamedTuple{base::Int}, typeof(string), Int))
           @eval f(i, base) = invoke(Core.kwcall, $m, (;base), string, i)
           end
    julia> f(20,16)
    "14"