Prevent pre-compilation target package from triggering extensions
It is possible for an extension `ExtAB` to be loadable by one of its
triggers, e.g. if A loads B. However this loading is only supposed
to happen after loading for A is finished, so it shouldn't be
included as part of pre-compiling A.

Getting this wrong means disagreeing with the scheduled pre-compile
jobs (A is not scheduled to depend on or generate a cache file for
ExtAB but accidentally does both) and leads to confusing errors
about missing cache files.

To avoid trying to use / generate a cache file for ExtAB while still
pre-compiling A, this change tracks the package being currently pre-
compiled so that its extension triggers can be ignored.