Add GC preserve hook (#70)
This PR ports https://github.com/mmtk/julia/pull/58 to `dev`. This PR is mostly the same as https://github.com/mmtk/julia/pull/58 except that 1. this PR does not remove transitive pinning of shadow stack roots (we know it is unsound to remove the transitive pinning at this stage), and 2. this PR includes minor refactoring for GC codegen interface.