Hide IRShow include from Revise
Revise in theory wants to re-evaluate this include, but it fails at doing so,
because the include call no longer works after bootstrap. It happens
to work right now on master, because the lowering of `Compiler.include`
happens to hide the include call from Revise, but that's a Revise bug
I'm about to fix. Address this by moving the include call into the
package and using an absolute include if necessary.