dict docs: document that ordering of keys/values/pairs match iterate (#56842)
Fix #56841.

Currently the documentation states that keys(dict) and values(dict)
iterate in the same order. But it is not stated whether this is the same
order as that used by pairs(dict), or when looping, for (k,v) in dict.

This PR makes this guarantee explicit.

(cherry picked from commit 796d823b4fbb3ec94c09f7022f9b614fc846dced)