Switch debug registry mapping to `CodeInstance`
Currently our debugging code maps from pointers to `MethodInstance`.
However, I think it makes more sense to map to `CodeInstance` instead,
since that's what we're actually compiling and as we're starting to
make more use of external owners and custom specializations, we'll
want to be able to annotate that in backtraces. This only adjusts
the internal data structures - any actual printing changes for
those sorts of use cases will have to come separately.