define `Core.Closure` and make compiler-generated closures subtype of it
Performance issues related to code containing closures are frequently
discussed (e.g., JuliaLang/julia#15276, JuliaLang/julia#56561). Several
approaches can be considered to address this problem, one of which
involves re-inferring code containing closures (JuliaLang/julia#56687).
To implement this idea, it is necessary to determine whether a given
piece of code includes a closure. However, there is currently no
explicit mechanism for making this determination (although there are
some code that checks whether the function name contains `"#"` for this
purpose, but this is an ad hoc solution).

To address this, this commit lays the foundation for future
optimizations targeting closures by defining closure functions as a
subtype of the new type `Core.Closure <: Function`. This change allows
the optimizer to apply targeted optimizations to code containing calls
to functions that are subtype of `Core.Closure`.