Local variable capture for methods of global functions
Here we introduce a new intermediate form `K"captured_local"` containing
an index into a list of the boxed local variables. After the `CodeInfo`
is constructed we then apply an extra rewrite step to replace any
`Expr(:captured_local)` with the associated boxed local variable.

Using a captured_local form allows us to finish lowering the lambda
expression (including linear IR construction) and eagerly construct a
CodeInfo, preserving the normal lowering workflow. In contrast, the
approach in the flisp code quoted the syntax for the lambda and ran
the macro `$` expansion pass again.