Add basic code for binding partition revalidation (#56649)
This adds the binding partition revalidation code from #54654. This is
the last piece of that PR that hasn't been merged yet - however the TODO
in that PR still stands for future work.

This PR itself adds a callback that gets triggered by deleting a
binding. It will then walk all code in the system and invalidate code
instances of Methods whose lowered source referenced the given global.
This walk is quite slow. Future work will add backedges and
optimizations to make this faster, but the basic functionality should be
in place with this PR.