Update HelloWorldC in Artifacts tests and bump LazyArtifacts for the same change (#57000)
This updates the version of HelloWorldC in
`stdlib/Artifacts/test/Artifacts.toml` from 1.1.2 to 1.4.0, for which
there are binaries for FreeBSD AArch64 and Linux RISC-V. The same change
was made to LazyArtifacts, as it contains that same `Artifacts.toml`
file in its tests, and the version of LazyArtifacts used here has been
updated accordingly. No other changes have been made to LazyArtifacts
since the previous version used here.