cleanup: Remove fallback post-lowering global resolution
After recent changes, essentially all global symbol scopes are
resolved by lowering, so the code in method.c that replaces
symbols by globalrefs is no longer necessary. The one small
exception to this were symbols resulting from closure conversion,
because these get inserted after the point at which lowering converts
symbols to globalrefs. However, in the current design, I think
that's basically a bug (and easily addressed by inserting
`globalref`s where appropriate). Removing this extra resolution
step is not particularly necessary, but for the upcoming
binding partition backedges, it removes an ambiguity as to
which version of the lowered code to scan. It also partially
resolves a very old todo about not performing post-hoc mutation
of lowered code (although we still do this for ccall).