Block thread from receiving profile signal with stackwalk lock
This is Jameson's proposed amendment to the changes made in 57035 that
introduced a deadlock on FreeBSD, amusingly in service of fixing a
deadlock on Linux.

On Linux and (non-macOS) BSD, instead of acquiring and releasing a lock
on the profiler in `jl_with_stackwalk_lock`, we're just blocking the
thread from receiving the profiler's `SIGUSR2` signal at all.

Should fix 57058. If so, closes 57059.

Co-Authored-By: Jameson Nash <vtjnash@gmail.com>