run UV on separate thread by default
This capability was added in https://github.com/JuliaLang/julia/pull/55529, but not used by default which seems silly since there are lots of issues like https://github.com/JuliaLang/julia/issues/43952 where this is needed to fix bugs. I'm not sure that this is the right place to add this (rather than in julia_init.c) but putting this up as a trial balloon.