Skip some stock GC-specific tests when using MMTk (#57104)
As noted in https://github.com/JuliaLang/julia/issues/57103, there are a
few tests that are specific to the stock GC and will fail for MMTk. For
now, this PR simply skips those tests when not using the stock GC until
we agree on how to approach the incompatibilities described in the
issue.