Add `K"SourceLocation"` for richer source location literals
Previously, we converted to LineNumberNode early in cases where the line
information needs to be emitted into the IR (eg, functionloc in method
signatures). However, this is lossy.

`K"SourceLocation"` is a "source provenance literal" - it doesn't hold
any data itself, but is merely a marker from which the source location
can later be extracted from the provenance tree.

Also factor out function argument matching of expand_function_def so it
can be called separately for opaque closure desugaring.