cfunction: reimplement, as originally planned, for reliable performance
This implements several sources of bugfixes and improvements:
 - direct edges are correctly represented
 - performance does not degrade when edges trigger
 - the JIT does not call `jl_infer_type` until actually required
 - constant return can handle invalidation and emitting efficient code

This generates the code for the exact signature specified by the user,
instead of using `jl_infer_type` to compute a different signature which
previously might cause unnecessary boxing and previously introduced
unstable performance characteristics. This lets us defer generating the
actual thunk required at runtime with the JIT when the required
information is already available, and also to validate that information
is still correct, and regenerate it when not correct anymore.